Rewrite data prep Jupyter notebook as a script
Implements the Jupyter notebook for data prep of US variants as a
Python script and collection of supporting data (clades to analyze and
state name/abbreviation mappings). We can run this script on demand or
through a scheduled job on GitHub. The updated README provides examples
of how to source from and push to S3 buckets. With a scheduled job, this
approach could allow us to fully automate the data preparation for US
variants except for occasional updates to the table of clades to
analyze.